refactor util - extract data gen
