+++ test/util/client_test.go
@@ -2,42 +2,13 @@ package util_test
 
 import (
 	"context"
-	"fmt"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
-	"github.com/solo-io/solo-kit/pkg/api/v1/resources/core"
-	"github.com/solo-io/squash/pkg/api/v1"
 	"github.com/solo-io/squash/pkg/options"
 	"github.com/solo-io/squash/test/util"
 )
-
-func generateDebugAttachment(name, namespace, dbgger, image, pod, container, processName string) v1.DebugAttachment {
-
-	da := v1.DebugAttachment{
-		Metadata: core.Metadata{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Debugger:  dbgger,
-		Image:     image,
-		Pod:       pod,
-		Container: container,
-	}
-	if processName != "" {
-		da.ProcessName = processName
-	}
-	return da
-}
-func generateDebugAttachmentDlv1(name, namespace string) v1.DebugAttachment {
-	dbgger := "dlv"
-	image := "mk"
-	pod := "somepod"
-	container := "somecontainer"
-	processName := "pcsnm"
-	return generateDebugAttachment(name, namespace, dbgger, image, pod, container, processName)
-}
 
 var _ = Describe("utils", func() {
 	It("should generate debug attachment", func() {
@@ -64,7 +35,6 @@ var _ = Describe("utils", func() {
 		Expect(err).To(BeNil())
 		Expect(read.Metadata.Name).To(Equal(name))
 		Expect(read.Metadata.Namespace).To(Equal(namespace))
-		fmt.Println(read)
 
 		// Cleanup
 		deleteOpts := clients.DeleteOpts{
